+++ DYJets/Includes/electrons.h
@@ -33,10 +33,10 @@ class electrons
     TTreeReaderArray<float> Electron_eta;
     TTreeReaderArray<float> Electron_phi;
     TTreeReaderArray<float> Electron_mass;
-    TTreeReaderArray<float> Electron_charge;
+    TTreeReaderArray<int> Electron_charge;
     TTreeReaderArray<float> Electron_deltaEtaSC;
     TTreeReaderArray<float> Electron_miniPFRelIso_all;
-    TTreeReaderArray<unsigned> Electron_cutBased;
+    TTreeReaderArray<int> Electron_cutBased;
 
     double _pt_cut = 20;
     double _eta_cut = 2.4;
